modularized functions added comments
