+++ award/data/namelist.csv
@@ -1,2 +1,2 @@
 Hname,Date,Cname,Type
-crap,"Nov 3rd, 2018",defaultUser,weekly
+pooping,"Nov 4th, 2018",defaultUser,weekly
